# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manajemen Kursus & Perencanaan API",
        "description": "L5 Swagger API documentation for Manajemen Kursus",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/post-test/{course}": {
            "get": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Pembuka post-test (informasi terkait pengerjaan post-test)",
                "description": "Menampilkan informasi sebelum memulai post-test",
                "operationId": "postTestInformation",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course Slug/ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informasi post-test",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post-test information"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Pengerjaan post-test",
                "description": "Mengirimkan jawaban post-test",
                "operationId": "submitPostTest",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course Slug/ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "answers"
                                ],
                                "properties": {
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "string",
                                                    "example": "uuid-question"
                                                },
                                                "answer_id": {
                                                    "type": "string",
                                                    "example": "uuid-answer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post-test berhasil disubmit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Submit post-test"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/post-test/result/{course}": {
            "get": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Detail hasil pengerjaan post-test",
                "description": "Melihat hasil pengerjaan post-test",
                "operationId": "postTestResult",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course Slug/ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil post-test",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post-test result"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pre-test/{course}": {
            "get": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Pembuka pre-test (informasi terkait pengerjaan pre-test)",
                "description": "Menampilkan informasi sebelum memulai pre-test",
                "operationId": "preTestInformation",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course Slug/ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informasi pre-test",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pre-test information"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Pengerjaan pre-test",
                "description": "Mengirimkan jawaban pre-test",
                "operationId": "submitPreTest",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course Slug/ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "answers"
                                ],
                                "properties": {
                                    "answers": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "question_id": {
                                                    "type": "string",
                                                    "example": "uuid-question"
                                                },
                                                "answer_id": {
                                                    "type": "string",
                                                    "example": "uuid-answer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pre-test berhasil disubmit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Submit pre-test"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pre-test/result/{course}": {
            "get": {
                "tags": [
                    "Guest Learning & Checkout"
                ],
                "summary": "Detail hasil pengerjaan pre-test",
                "description": "Melihat hasil pengerjaan pre-test",
                "operationId": "preTestResult",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "Course Slug/ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hasil pre-test",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Pre-test result"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}